Start working up a generic pub/sub publisher service.
